Restructure the component folder
