+++ Admin/src/components/sharedLayouts/adminSidebar/adminSideBar.tsx
@@ -5,17 +5,10 @@ import styles from "./sideBar.module.css";
 import Image from "next/image";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { useRouter } from "next/navigation";
-import { useSearchParams } from "next/navigation";
 import { adminSideBar } from "@/data/datarafce";
 
 const SharedAdminSideBar = () => {
   const pathname = usePathname();
-  const router = useRouter();
-
-  const searchParam = useSearchParams();
-
-  // console.log(router.push);
 
   return (
     <section className={styles.admin__side__bar}>
