+++ Admin/src/types/index.ts
@@ -17,7 +17,7 @@ export interface InputProps {
   type: string;
   name: string;
   holder: string;
-  icons?: string | StaticImport;
+  icons?: any;
   value: string;
   handleChange: (e: ChangeEvent<HTMLInputElement>) => void;
   inputStyle?: string;
